角色區塊拆分完畢，進一步優化之後再搞 markdown 怎麼在 vue 元件生效找到了，但搞了兩小時沒反應，草
